HW2/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
abseil-cpp=20211102.0=h96cf925_1
aiohttp=3.8.3=py39ha30fb19_0
aiosignal=1.2.0=pyhd8ed1ab_0
arrow-cpp=8.0.0=py39had1886b_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.1.0=pyh71513ae_1
autopep8=1.6.0=pyhd3eb1b0_1
aws-c-common=0.4.57=hb1e8313_1
aws-c-event-stream=0.1.6=h23ab428_5
aws-checksums=0.1.9=hb1e8313_0
aws-sdk-cpp=1.8.185=he271ece_0
blas=1.0=mkl
boost-cpp=1.70.0=hd59e818_1
brotli=1.0.9=h5eb16cf_7
brotli-bin=1.0.9=h5eb16cf_7
brotlipy=0.7.0=py39h9ed2024_1003
bzip2=1.0.8=h1de35cc_0
c-ares=1.18.1=h0d85af4_0
ca-certificates=2022.9.24=h033912b_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py39hc55c11b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39hecd8cb5_0
colorama=0.4.5=py39hecd8cb5_0
cryptography=37.0.1=py39hf6deb26_0
dataclasses=0.8=pyh6d0b6a4_7
datasets=2.6.1=py_0
dill=0.3.5.1=pyhd8ed1ab_0
ffmpeg=4.2.2=h97e5cf8_0
filelock=3.6.0=pyhd3eb1b0_0
freetype=2.11.0=hd8bbffd_0
frozenlist=1.3.1=py39h6218fd2_0
fsspec=2022.8.2=pyhd8ed1ab_0
gettext=0.21.0=h7535e17_0
gflags=2.2.2=hb1e8313_1004
giflib=5.2.1=haf1e3a3_0
glog=0.6.0=h8ac2a54_0
gmp=6.2.1=he9d5cce_3
gnutls=3.6.15=hed9c0bf_0
grpc-cpp=1.46.1=h067a048_0
huggingface_hub=0.10.1=py_0
icu=58.2=h0a44026_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py39hecd8cb5_0
importlib_metadata=4.11.3=hd3eb1b0_0
intel-openmp=2021.4.0=hecd8cb5_3538
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=hca72f7f_0
krb5=1.19.3=hb49756b_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libbrotlicommon=1.0.9=h5eb16cf_7
libbrotlidec=1.0.9=h5eb16cf_7
libbrotlienc=1.0.9=h5eb16cf_7
libcurl=7.84.0=h6dfd666_0
libcxx=14.0.6=h9765a3e_0
libdeflate=1.8=h9ed2024_5
libedit=3.1.20191231=h0678c8f_2
libev=4.33=haf1e3a3_1
libevent=2.1.10=h815e4d9_4
libffi=3.3=hb1e8313_2
libiconv=1.16=hca72f7f_2
libidn2=2.3.2=h9ed2024_0
libnghttp2=1.46.0=ha29bfda_0
libopus=1.3.1=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libprotobuf=3.20.1=h8346a28_0
libssh2=1.10.0=h52ee1ee_2
libtasn1=4.16.0=h9ed2024_0
libthrift=0.15.0=h054ceb0_0
libtiff=4.4.0=h2ef1027_0
libunistring=0.9.10=h9ed2024_0
libvpx=1.7.0=h378b8a2_0
libwebp=1.2.4=h56c3ce4_0
libwebp-base=1.2.4=hca72f7f_0
libxml2=2.9.14=hbf8cd5e_0
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.3=h23ab428_1
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.1=py39h4ab4a9b_0
mkl_random=1.2.2=py39hb2f4e1b_0
multidict=6.0.2=py39h63b48b0_1
multiprocess=0.70.13=py39h701faf5_1
ncurses=6.3=hca72f7f_3
nettle=3.7.3=h230ac6f_1
numpy=1.23.1=py39h2e5f0a9_0
numpy-base=1.23.1=py39h3b1a694_0
openh264=2.1.1=h8346a28_0
openssl=1.1.1q=hfe4f2af_0
orc=1.7.4=h9274d09_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.5.0=py39hecff1ad_0
pillow=9.2.0=py39hde71d04_1
pip=22.2.2=py39hecd8cb5_0
protobuf=3.20.1=py39he9d5cce_0
pyarrow=8.0.0=py39h2202ef3_0
pycodestyle=2.8.0=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py39hecd8cb5_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.13=hdfd78df_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python-xxhash=3.0.0=py39h63b48b0_1
python_abi=3.9=2_cp39
pytorch=1.13.0.dev20221002=py3.9_0
pytz=2022.4=pyhd8ed1ab_0
pyyaml=6.0=py39hca72f7f_1
re2=2022.04.01=h96cf925_0
readline=8.1.2=hca72f7f_1
regex=2022.7.9=py39hca72f7f_0
requests=2.28.1=py39hecd8cb5_0
sacremoses=master=py_0
setuptools=63.4.1=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h6e38e02_1
sqlite=3.39.3=h707629a_0
tk=8.6.12=h5d9f67b_0
tokenizers=0.13.0.dev0=py39_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.13.0.dev20221002=py39_cpu
torchvision=0.14.0.dev20221002=py39_cpu
tqdm=4.64.1=pyhd8ed1ab_0
transformers=4.23.1=py_0
typing-extensions=4.3.0=py39hecd8cb5_0
typing_extensions=4.3.0=py39hecd8cb5_0
tzdata=2022c=h04d1e81_0
urllib3=1.26.11=py39hecd8cb5_0
utf8proc=2.6.1=h9ed2024_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!157.20191217=h1de35cc_0
xxhash=0.8.0=h35c211d_3
xz=5.2.6=hca72f7f_0
yaml=0.2.5=haf1e3a3_0
yarl=1.7.2=py39h63b48b0_2
zipp=3.8.0=py39hecd8cb5_0
zlib=1.2.12=h4dc903c_3
zstd=1.5.2=hcb37349_0
